+++ gen_tests/unsupported_auth/lib/api.dart
@@ -1,0 +1,3 @@
+export 'package:unsupported_auth/api/default_api.dart';
+export 'package:unsupported_auth/api_client.dart';
+export 'package:unsupported_auth/api_exception.dart';