+++ gen_tests/unsupported_auth/lib/client.dart
@@ -1,0 +1,9 @@
+import 'package:unsupported_auth/api.dart';
+
+class Unsupported {
+  Unsupported({ApiClient? client}) : client = client ?? ApiClient();
+
+  final ApiClient client;
+
+  DefaultApi get defaultApi => DefaultApi(client);
+}